# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workflows": {
      "post": {
        "operationId": "WorkflowsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "WorkflowsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{id}": {
      "put": {
        "operationId": "WorkflowsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "WorkflowsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "workflows"
        ]
      },
      "delete": {
        "operationId": "WorkflowsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{id}/commit": {
      "post": {
        "operationId": "WorkflowsController_commit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compiled workflow definition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "entrypoint": {
                      "type": "object",
                      "properties": {
                        "ref": {
                          "type": "string"
                        }
                      }
                    },
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "ref": {
                            "type": "string"
                          },
                          "componentId": {
                            "type": "string"
                          },
                          "params": {
                            "type": "object",
                            "additionalProperties": true
                          },
                          "dependsOn": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "config": {
                      "type": "object",
                      "properties": {
                        "environment": {
                          "type": "string"
                        },
                        "timeoutSeconds": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{id}/run": {
      "post": {
        "operationId": "WorkflowsController_run",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Workflow execution result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string",
                      "description": "Temporal workflow identifier"
                    },
                    "workflowId": {
                      "type": "string",
                      "description": "Workflow record id"
                    },
                    "temporalRunId": {
                      "type": "string",
                      "description": "Temporal first execution run id"
                    },
                    "taskQueue": {
                      "type": "string",
                      "description": "Temporal task queue used for execution"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "RUNNING",
                        "COMPLETED",
                        "FAILED",
                        "CANCELLED",
                        "TERMINATED",
                        "CONTINUED_AS_NEW",
                        "TIMED_OUT",
                        "UNKNOWN"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/runs/{runId}/status": {
      "get": {
        "operationId": "WorkflowsController_status",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "temporalRunId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current Temporal execution status"
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/runs/{runId}/result": {
      "get": {
        "operationId": "WorkflowsController_result",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "temporalRunId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved workflow result payload"
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/runs/{runId}/cancel": {
      "post": {
        "operationId": "WorkflowsController_cancel",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "temporalRunId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancels a running workflow execution"
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/runs/{runId}/trace": {
      "get": {
        "operationId": "WorkflowsController_trace",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trace events for a workflow run",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string"
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string"
                          },
                          "nodeRef": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "message": {
                            "type": "string"
                          },
                          "error": {
                            "type": "string"
                          },
                          "outputSummary": {
                            "type": "object"
                          }
                        },
                        "additionalProperties": false
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
...
  },
  "info": {
    "title": "ShipSec Studio API",
    "description": "ShipSec backend API specification",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateWorkflowRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "nodes": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "position": {
                  "type": "object",
                  "properties": {
                    "x": {
                      "type": "number"
                    },
                    "y": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "x",
                    "y"
                  ]
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "label": {
                      "type": "string"
                    },
                    "config": {
                      "default": {},
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "label"
                  ]
                }
              },
              "required": [
                "id",
                "type",
                "position",
                "data"
              ]
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "source": {
                  "type": "string"
                },
                "target": {
                  "type": "string"
                },
                "sourceHandle": {
                  "type": "string"
                },
                "targetHandle": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "source",
                "target"
              ]
            }
          },
          "viewport": {
            "type": "object",
            "properties": {
              "x": {
                "type": "number"
              },
              "y": {
                "type": "number"
              },
              "zoom": {
                "type": "number"
              }
            },
            "required": [
              "x",
              "y",
              "zoom"
            ]
          }
        },
        "required": [
          "name",
          "nodes",
          "edges",
          "viewport"
        ]
      },
      "UpdateWorkflowRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "nodes": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "position": {
                  "type": "object",
                  "properties": {
                    "x": {
                      "type": "number"
                    },
                    "y": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "x",
                    "y"
                  ]
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "label": {
                      "type": "string"
                    },
                    "config": {
                      "default": {},
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "label"
                  ]
                }
              },
              "required": [
                "id",
                "type",
                "position",
                "data"
              ]
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "source": {
                  "type": "string"
                },
                "target": {
                  "type": "string"
                },
                "sourceHandle": {
                  "type": "string"
                },
                "targetHandle": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "source",
                "target"
              ]
            }
          },
          "viewport": {
            "type": "object",
            "properties": {
              "x": {
                "type": "number"
              },
              "y": {
                "type": "number"
              },
              "zoom": {
                "type": "number"
              }
            },
            "required": [
              "x",
              "y",
              "zoom"
            ]
          }
        },
        "required": [
          "name",
          "nodes",
          "edges",
          "viewport"
        ]
      }
    }
  }
}