# backend/src/workflows/dto/workflow-graph.dto.ts
import { createZodDto } from 'nestjs-zod';
import { z } from 'zod';

export const WorkflowViewportSchema = z.object({
  x: z.number(),
  y: z.number(),
  zoom: z.number(),
});

export const WorkflowNodeDataSchema = z.object({
  label: z.string(),
  config: z.record(z.string(), z.unknown()).default({}),
});

export const WorkflowNodeSchema = z.object({
  id: z.string(),
  type: z.string(),
  position: z.object({
    x: z.number(),
    y: z.number(),
  }),
  data: WorkflowNodeDataSchema,
});

export const WorkflowEdgeSchema = z.object({
  id: z.string(),
  source: z.string(),
  target: z.string(),
  sourceHandle: z.string().optional(),
  targetHandle: z.string().optional(),
});

export const WorkflowGraphSchema = z.object({
  id: z.string().optional(),
  name: z.string(),
  description: z.string().optional(),
  nodes: z.array(WorkflowNodeSchema).min(1),
  edges: z.array(WorkflowEdgeSchema),
  viewport: WorkflowViewportSchema,
});

export type WorkflowGraph = z.infer<typeof WorkflowGraphSchema>;

export class WorkflowGraphDto extends createZodDto(WorkflowGraphSchema) {}
export class CreateWorkflowRequestDto extends WorkflowGraphDto {}
export class UpdateWorkflowRequestDto extends WorkflowGraphDto {}